Mostly restore inline replace extension tests
